extract InputManager
